# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Estudiantes",
        "description": "Documentación de la API para gestionar estudiantes",
        "contact": {
            "email": "soporte@tuapi.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/students": {
            "get": {
                "tags": [
                    "Estudiantes"
                ],
                "summary": "Obtiene la lista de todos los estudiantes",
                "operationId": "a49cdf9c0f31741274dcf5e369bf723c",
                "responses": {
                    "200": {
                        "description": "Lista de estudiantes obtenida exitosamente"
                    },
                    "404": {
                        "description": "No se encontraron estudiantes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Estudiantes"
                ],
                "summary": "Registra un nuevo estudiante",
                "operationId": "dee128cbe5cb07a7ec73a98af9d49dbc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone",
                                    "language"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "language": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Estudiante creado exitosamente"
                    },
                    "400": {
                        "description": "Error en la validación de datos"
                    }
                }
            }
        },
...
    },
...
}